Clean up folder structure.
