+++ redez-sem-hs20/groups/03-doubleRatchet/src/helpers/signing.py
@@ -1,5 +1,5 @@
 from xeddsa.implementations import XEdDSA25519
-from helpers import FOLDERNAME_KEYS
+from helper_functions.helpers import FOLDERNAME_KEYS
 import os
 
 # This is not cryptographically safe. To be that the program should be a fixed set of operations that run every time.
